refactor(step2): Players 클래스 생성
